chore: LecueBook type 분리
